Add vec tests
